CustomCode & ASGNode Update
Deleting a connection deletes it from DB as well, and removes the deleted node from in- and out-connection lists. Adding a link (not yet perfect) adds it to the DB if not previously present, and populates it's in- and out-connection lists, as well as those of the nodes connected to it.